moving 2.0.0 from folder to main, deprecated <1.0.0 files
