+++ .gitignore
@@ -1,1 +1,2 @@
 node_modules/
+require/_file_cache.json
