+++ index.html
@@ -1,26 +1,24 @@
 <!DOCTYPE html>
 <html>
+    <head>
+        <meta charset="UTF-8">
+        <title>Electron - Life Engine</title>
+        <!-- You can also require other files to run in this process -->
+        
+        <script src="./js/app.js" type="text/javascript" charset="utf-8"></script>
+        <style>
+            body {
+                margin: 0;
+                padding: 0;
+                -webkit-app-region: drag;
+                -webkit-user-select: none;
+            }
 
-<head>
-  <meta charset="UTF-8">
-  <!-- You can also require other files to run in this process -->
-  <script src="./js/src/sketch.js" type="text/javascript" charset="utf-8"></script>
-  <style>
-    body {
-      margin: 0;
-      padding: 0;
-      -webkit-app-region: drag;
-      -webkit-user-select: none;
-    }
-
-    canvas {
-      display: block;
-    }
-  </style>
-  <title>P5 Electron Quick Start</title>
-</head>
-
-<body>
-</body>
-
+            canvas {
+                display: block;
+            }
+        </style>
+    </head>
+    <body>
+    </body>
 </html>
