+++ js-deprecated/interactions/crystal.hjson
