+++ js-deprecated/interactions/deprecated/Electron.hjson
