+++ js-deprecated/interactions/deprecated/lattice.hjson
