+++ js-deprecated/interactions/generateTypesObject.js
