+++ js-deprecated/interactions/parasite.hjson
