+++ js-deprecated/src/Particle.js
