+++ js-deprecated/src/camera.js
