+++ js-deprecated/src/mouse.js
