+++ js-deprecated/utils/require_hjson.js
