+++ js-deprecated/utils/toProperCase.js
