+++ require/require-extensions.js
