+++ ui/js/DualSlider.js
